host/requirements.txt
pydantic_settings>=2.5.2,<2.6.0
pyserial>=3.5.0,<3.6.0
requests>=2.32.3,<2.33.0